refactor: Rename agent.types.ts to swarm.types.ts and update imports
